+++ feature/account/domain/api/src/main/java/ru/crazerr/feature/account/domain/api/Account.kt
@@ -3,9 +3,9 @@ package ru.crazerr.feature.account.domain.api
 import ru.crazerr.feature.currency.domain.api.Currency
 
 data class Account(
-    val id: Int,
+    val id: Long,
     val name: String,
-    val amount: Long,
+    val amount: Double,
     val iconId: String,
     val currency: Currency,
 )
