+++ feature/account/presentation/src/main/java/ru/crazerr/feature/account/presentation/AccountEditorArgs.kt
@@ -1,5 +1,5 @@
 package ru.crazerr.feature.account.presentation
 
 data class AccountEditorArgs(
-    val accountId: Int = -1,
+    val accountId: Long = -1,
 )
