+++ feature/budget/domain/src/main/java/ru/crazerr/feature/budget/domain/BudgetRepository.kt
@@ -3,7 +3,7 @@ package ru.crazerr.feature.budget.domain
 import ru.crazerr.feature.domain.api.Budget
 
 interface BudgetRepository {
-    suspend fun getBudgetById(id: Int): Result<Budget>
+    suspend fun getBudgetById(id: Long): Result<Budget>
 
     suspend fun createBudget(budget: Budget): Result<Budget>
