+++ feature/budget/domain/src/main/java/ru/crazerr/feature/budget/domain/TransactionRepository.kt
@@ -3,5 +3,5 @@ package ru.crazerr.feature.budget.domain
 import java.time.LocalDate
 
 interface TransactionRepository {
-    suspend fun getCurrentAmount(categoryId: Int, date: LocalDate): Result<Long>
+    suspend fun getCurrentAmount(categoryId: Long, date: LocalDate): Result<Double>
 }
