+++ feature/budget/presentation/src/main/java/ru/crazerr/feature/budget/presentation/budgetStory/BudgetStoryArgs.kt
@@ -1,5 +1,5 @@
 package ru.crazerr.feature.budget.presentation.budgetStory
 
 class BudgetStoryArgs(
-    val budgetId: Int,
+    val budgetId: Long,
 )
