+++ feature/category/domain/src/main/java/ru/crazerr/feature/category/domain/repository/CategoryRepository.kt
@@ -3,7 +3,7 @@ package ru.crazerr.feature.category.domain.repository
 import ru.crazerr.feature.domain.api.Category
 
 interface CategoryRepository {
-    suspend fun getCategoryById(id: Int): Result<Category>
+    suspend fun getCategoryById(id: Long): Result<Category>
 
     suspend fun createCategory(category: Category): Result<Category>
