+++ feature/category/presentation/src/main/java/ru/crazerr/feature/category/presentation/categoryEditor/CategoryEditorArgs.kt
@@ -1,5 +1,5 @@
 package ru.crazerr.feature.category.presentation.categoryEditor
 
 data class CategoryEditorArgs(
-    val id: Int = -1,
+    val id: Long = -1,
 )
