+++ feature/main/presentation/src/main/java/ru/crazerr/feature/main/presentation/main/MainComponentAction.kt
@@ -1,5 +1,5 @@
 package ru.crazerr.feature.main.presentation.main
 
 sealed interface MainComponentAction {
-    data class GoToAccount(val id: Int = -1) : MainComponentAction
+    data class GoToAccount(val id: Long = -1) : MainComponentAction
 }
