+++ feature/main/presentation/src/main/java/ru/crazerr/feature/main/presentation/main/MainViewAction.kt
@@ -4,5 +4,5 @@ import ru.crazerr.feature.account.domain.api.Account
 
 sealed interface MainViewAction {
     data class DeleteAccount(val account: Account, val index: Int) : MainViewAction
-    data class GoToAccount(val id: Int = -1) : MainViewAction
+    data class GoToAccount(val id: Long = -1) : MainViewAction
 }
