+++ feature/transactions/presentation/src/main/java/ru/crazerr/feature/transactions/presentation/transactionsFilter/TransactionsFilterArgs.kt
@@ -3,8 +3,8 @@ package ru.crazerr.feature.transactions.presentation.transactionsFilter
 import java.time.LocalDate
 
 class TransactionsFilterArgs(
-    val accountIds: IntArray,
-    val categoryIds: IntArray,
+    val accountIds: LongArray,
+    val categoryIds: LongArray,
     val startDate: LocalDate?,
     val endDate: LocalDate?,
     val isFilterEnabled: Boolean,
