+++ feature/transactions/presentation/src/main/java/ru/crazerr/feature/transactions/presentation/transactionsFilter/TransactionsFilterComponentAction.kt
@@ -5,8 +5,8 @@ import java.time.LocalDate
 sealed interface TransactionsFilterComponentAction {
     data object BackClick : TransactionsFilterComponentAction
     data class SaveButtonClick(
-        val accountIds: IntArray,
-        val categoryIds: IntArray,
+        val accountIds: LongArray,
+        val categoryIds: LongArray,
         val startDate: LocalDate?,
         val endDate: LocalDate?,
         val isFilterEnabled: Boolean
